RavenDB-19208 Create indexing configuration to switch between VoronInputIndex and VoronBufferedInput
