+++ src/Raven.Client/Documents/Indexes/LuceneIndexInputType.cs
@@ -1,0 +1,7 @@
+﻿namespace Raven.Client.Documents.Indexes;
+
+public enum LuceneIndexInputType
+{
+    Standard,
+    Buffered
+}
