+++ test/FastTests/Issues/RavenDB_16590.cs
@@ -75,6 +75,7 @@ namespace FastTests.Issues
                 "Indexing.TransactionSizeLimitInMb",
                 "Indexing.OrderByScoreAutomaticallyWhenBoostingIsInvolved",
                 "Indexing.Lucene.UseCompoundFileInMerging",
+                "Indexing.Lucene.IndexInputType"
 
                 
                 //Obsolete studio keys:
